requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
pyodbc==4.0.39
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
Pillow==10.0.0
matplotlib==3.7.2
gunicorn==21.2.0
requests==2.31.0